+++ src/Database/Design/Ampersand/Prototype/Generate.hs
@@ -28,14 +28,6 @@ generateDBstructQueries fSpec withComment
         , "   , "<>doubleQuote "lastAccess"<>" BIGINT NOT NULL"
         , "   ) ENGINE="<>dbEngine
         ]
-      , [ "CREATE TABLE "<> doubleQuote "__History__"
-        , "   ( "<>doubleQuote "Seconds"<>" VARCHAR(255) DEFAULT NULL"
-        , "   , "<>doubleQuote "Date"<>" VARCHAR(255) DEFAULT NULL"
-        , "   ) ENGINE="<>dbEngine
-        ]
-      , [ "INSERT INTO "<>doubleQuote "__History__"<>" ("<>doubleQuote "Seconds"<>","<>doubleQuote "Date"<>")"
-        , "   VALUES (UNIX_TIMESTAMP(NOW(6)), NOW(6))"
-        ]
       , [ "CREATE TABLE "<> doubleQuote "__all_signals__"
         , "   ( "<>doubleQuote "conjId"<>" VARCHAR(255) NOT NULL"
         , "   , "<>doubleQuote "src"<>" VARCHAR(255) NOT NULL"
@@ -104,7 +96,10 @@ plug2TableSpec plug
      , tsName = name plug
      , tsflds = plugAttributes plug
      , tsKey  = case (plug, (head.plugAttributes) plug) of
-                 (BinSQL{}, _)   -> []
+                 (BinSQL{}, _)   -> [  "PRIMARY KEY (" 
+                                       <> intercalate ", " (map (show . attName) (plugAttributes plug))
+                                       <> ")"
+                                    ]
                  (TblSQL{}, primFld) ->
                       case attUse primFld of
                          PrimaryKey _ -> ["PRIMARY KEY (" <> (show . attName) primFld <> ")" ]
@@ -127,7 +122,7 @@ generateAllDefPopQueries fSpec
     fillSignalTable :: [(Conjunct, [AAtomPair])] -> [Text.Text]
     fillSignalTable [] = []
     fillSignalTable conjSignals 
-     = [Text.unlines $ 
+     = [Text.unlines
             [ "INSERT INTO "<>Text.pack (show (getTableName signalTableSpec))
             , "   ("<>Text.intercalate ", " (map (Text.pack . doubleQuote) ["conjId","src","tgt"])<>")"
             , "VALUES " <> Text.intercalate " , " 
@@ -146,7 +141,7 @@ generateAllDefPopQueries fSpec
           = case tableContents fSpec plug of
              []  -> []
              tblRecords 
-                 -> [Text.unlines $ 
+                 -> [Text.unlines
                        [ "INSERT INTO "<>Text.pack (show (name plug))
                        , "   ("<>Text.intercalate ", " (map (Text.pack . show . attName) (plugAttributes plug))<>") "
                        , "VALUES " <> Text.intercalate " , "