+++ static/zwolle/src/Ampersand/Interfacing/InterfaceObject.php
@@ -59,10 +59,10 @@ class InterfaceObject {
 	public $label;
 	
 	/**
-	 * Specifies is this interface object is a toplevel interface (true) or subinterface (false)
+	 * Specifies if this interface object is a toplevel interface (true) or subinterface (false)
 	 * @var boolean
 	 */
-	public $isTopLevelIfc = false;
+	private $isRoot;
 	
 	/**
 	 * Roles that have access to this interface
@@ -189,17 +189,19 @@ class InterfaceObject {
 
 	/**
 	 * InterfaceObject constructor
-	 * @param array $ifcDef
+	 * @param array $ifcDef Interface object definition as provided by Ampersand generator
 	 * @param string $pathEntry
+     * @param boolean $rootIfc Specifies if this interface object is a toplevel interface (true) or subinterface (false)
 	 */
-	public function __construct($ifcDef, $pathEntry = null){
+	public function __construct($ifcDef, $pathEntry = null, $rootIfc = false){
 		$this->database = Database::singleton();
 		$this->logger = Logger::getLogger('FW');
 		
+        $this->isRoot = $rootIfc;
+        
 		// Set attributes from $ifcDef
 		$this->id = $ifcDef['id'];
 		$this->label = $ifcDef['label'];
-		
 		$this->view = is_null($ifcDef['viewId']) ? null : View::getView($ifcDef['viewId']);
 		
 		$this->path = is_null($pathEntry) ? $this->id : "{$pathEntry}/{$this->id}";
@@ -260,7 +262,16 @@ class InterfaceObject {
 	public function __toString() {
 		return $this->id;
 	}
-	
+    
+    /**
+     * Returns interface relation (when interface expression = relation), throws exception otherwise
+     * @return Relation|Exception
+     */
+	public function relation(){
+        if(is_null($this->relation)) throw new Exception ("Interface expression for '{$this->path}' is not a relation", 500);
+        else return $this->relation;
+    }
+
 	/**
 	 * Returns if interface expression relation is a property
 	 * @return boolean
@@ -276,7 +287,23 @@ class InterfaceObject {
     public function isRef(){
         return !is_null($this->refInterfaceId);
     }
-	
+    
+    /**
+     * Returns if interface object is a top level interface
+     * @return boolean
+     */
+    public function isRoot(){
+        return $this->isRoot;
+    }
+    
+    /**
+     * Returns if interface is a public interface (i.e. accessible every role, incl. no role)
+	 * @return boolean
+     */
+    public function isPublic(){
+        return empty($this->ifcRoleNames) && $this->isRoot();
+    }
+    
 	/**
 	 * 
 	 * @param Atom $atom
@@ -302,10 +329,28 @@ class InterfaceObject {
 	    
 	    throw new Exception("Subinterface '{$ifcLabel}' does not exists in interface '{$this->path}'", 500);
 	}
+    
+    /**
+     * Return array with all sub interface recursively (incl. the interface itself)
+     * @return InterfaceObject[]
+     */
+    public function getInterfaceFlattened(){
+        $arr = array();
+        $arr[] = $this;
+        foreach ($this->getSubinterfaces(false) as $ifc){
+            $arr = array_merge($arr, $ifc->getInterfaceFlattened());
+        }
+        return $arr;
+    }
 	
-	private function getSubinterfaces(){
+    /**
+     * @param boolean $inclRefs specifies whether to include subinterfaces from referenced interfaces
+     * @return InterfaceObject[] 
+     */
+	private function getSubinterfaces($inclRefs = true){
 	    if(!$this->isRef()) return $this->subInterfaces;
-	    else return self::getInterface($this->refInterfaceId)->getSubinterfaces();
+	    elseif($inclRefs) return self::getInterface($this->refInterfaceId)->getSubinterfaces();
+        else return array();
 	}
 	
 /**************************************************************************************************
@@ -331,7 +376,7 @@ class InterfaceObject {
 	    // Check if atom does not exist and if it may be created here
 	    elseif(!$atom->atomExists() && $this->crudC){
 	        // If interface expression is a relation, add tuple($this->srcAtom, $atom) in this relation
-	        if($this->relation) $this->relation->addLink($this->srcAtom, $atom, $this->relationIsFlipped);
+	        if($this->relation) $this->relation()->addLink($this->srcAtom, $atom, $this->relationIsFlipped);
 	        // Else only create atom
 	        else $atom->addAtom();
 	        
@@ -410,13 +455,8 @@ class InterfaceObject {
 	 * @return mixed
 	 */
 	public function getContent($options = array(), $recursionArr = array(), $depth = null){
-	    // CRUD check
-	    if(!$this->crudR) throw new Exception("Read not allowed for '{$this->path}'", 405);
-	    
-	    // Default options
+        // Default options
 	    $options['arrayType'] = isset($options['arrayType']) ? $options['arrayType'] : 'num';
-	    $options['metaData'] = isset($options['metaData']) ? filter_var($options['metaData'], FILTER_VALIDATE_BOOLEAN) : true;
-	    $options['navIfc'] = isset($options['navIfc']) ? filter_var($options['navIfc'], FILTER_VALIDATE_BOOLEAN) : true;
 	    $options['inclLinktoData'] = isset($options['inclLinktoData']) ? filter_var($options['inclLinktoData'], FILTER_VALIDATE_BOOLEAN) : false;
         if(isset($options['depth']) && is_null($depth)) $depth = $options['depth']; // initialize depth, if specified in options array
 	    
@@ -507,28 +547,44 @@ class InterfaceObject {
 	
 /**************************************************************************************************
  *
- * CREATE, UPDATE, PATCH and DELETE functions
+ * READ, CREATE, UPDATE, PATCH and DELETE functions
  *
  *************************************************************************************************/
-	
-	/**
+    
+    /**
+    * @param array $options 
+    * @throws Exception when read is not allowed for this interface object
+    * @return mixed
+    */
+    public function read($options = []){
+        $this->logger->debug("read() called on {$this->path}");
+        
+        // CRUD check
+        if(!$this->crudR) throw new Exception("Read not allowed for '{$this->path}'", 405);
+        
+        return $this->getContent($options);
+    }
+    
+    /**
 	 * Function to create a new Atom at the given interface.
 	 * @param array $data
 	 * @param array $options
 	 * @throws Exception
 	 * @return mixed
 	 */
-	public function create($data, $options = array()){	
+	public function create($data, $options = array()){
+        $this->logger->debug("create() called on {$this->path}");
+        
 	    // CRUD check
 	    if(!$this->crudC) throw new Exception ("Create not allowed for '{$this->path}'", 405);
 	    if(!$this->tgtConcept->isObject) throw new Exception ("Cannot create non-object '{$this->tgtConcept}' in '{$this->path}'. Use PATCH add operation instead", 405);
+        if($this->isRef()) throw new Exception ("Cannot create on reference interface in '{$this->path}'. See #498", 501);
 	    
 	    // Handle options
 	    if(isset($options['requestType'])) $this->database->setRequestType($options['requestType']);
 	
 	    // Perform create
 	    $newAtom = $this->tgtConcept->createNewAtom();
-	    $newAtom->addAtom();
 	
 	    // Special case for CREATE in I[Concept] interfaces
 	    if($this->srcAtom->id === '_NEW'){
@@ -537,8 +593,9 @@ class InterfaceObject {
 	    }
 	
 	    // If interface expression is a relation, also add tuple(this, newAtom) in this relation
-	    if($this->relation) $this->relation->addLink($this->srcAtom, $newAtom, $this->relationIsFlipped);
-	    
+	    if($this->relation) $this->relation()->addLink($this->srcAtom, $newAtom, $this->relationIsFlipped);
+        else $newAtom->addAtom();
+        
 	    // Walk to new atom
 	    $newAtom = $this->atom($newAtom->id);
 	    
@@ -574,12 +631,10 @@ class InterfaceObject {
 	    }
 	
 	    // Close transaction
-	    $atomStoreNewContent = $this->crudR ? $newAtom : null; // Get and store new content if interface is readable (crudR)
-	    $this->database->closeTransaction($newAtom->concept . ' created', null, $atomStoreNewContent);
+	    $this->database->closeTransaction($newAtom->concept . ' created', null, $newAtom); // temp store content of $newAtom (also when not crudR)
 	
 	    // Return atom content (can be null)
 	    return $newAtom->getStoredContent();
-	
 	}
 	
 	/**
@@ -587,6 +642,7 @@ class InterfaceObject {
 	 * @throws Exception
 	 */
 	public function update(){
+        $this->logger->debug("update() called on {$this->path}");
 	    throw new Exception ("Cannot update from interface '{$this->path}'. Add resource identifier behind path", 405);
 	}
 	
@@ -595,6 +651,7 @@ class InterfaceObject {
 	 * @throws Exception
 	 */
 	public function patch(){
+        $this->logger->debug("patch() called on {$this->path}");
 	    throw new Exception ("Cannot patch from interface '{$this->path}'. Add resource identifier behind path", 405);
 	}
 	
@@ -603,6 +660,7 @@ class InterfaceObject {
 	 * @throws Exception
 	 */
 	public function delete(){
+        $this->logger->debug("delete() called on {$this->path}");
 	    throw new Exception ("Cannot delete from interface '{$this->path}'. Add resource identifier behind path", 405);
 	}
 	
@@ -621,6 +679,7 @@ class InterfaceObject {
 	public function doPatchReplace($patch){
 	    // CRUD check
 	    if(!$this->crudU) throw new Exception("Update is not allowed for path '{$this->path}'", 403);
+        if($this->isRef()) throw new Exception ("Cannot update on reference interface in '{$this->path}'. See #498", 501);
 	
 	    // PatchReplace only works for UNI expressions. Otherwise, use patch remove and patch add
 	    if(!$this->isUni) throw new Exception("Cannot patch replace for non-univalent interface '{$this->path}'. Use patch remove + add instead", 500);
@@ -635,9 +694,9 @@ class InterfaceObject {
 	        if(!(is_bool($value) || is_null($value))) throw new Exception("Interface '{$this->path}' is property, boolean expected, non-boolean provided");
 	        	
 	        // When true
-	        if($value) $this->relation->addLink($this->srcAtom, $this->srcAtom, $this->relationIsFlipped);
+	        if($value) $this->relation()->addLink($this->srcAtom, $this->srcAtom, $this->relationIsFlipped);
 	        // When false or null
-	        else $this->relation->deleteLink($this->srcAtom, $this->srcAtom, $this->relationIsFlipped);
+	        else $this->relation()->deleteLink($this->srcAtom, $this->srcAtom, $this->relationIsFlipped);
 	        	
 	    // Interface is a relation to an object
 	    }elseif($this->tgtConcept->isObject){
@@ -645,11 +704,10 @@ class InterfaceObject {
 	
 	    // Interface is a relation to a scalar (i.e. not an object)
 	    }elseif(!$this->tgtConcept->isObject){
-	        
 	        // Replace by nothing => deleteLink
-	        if(is_null($value)) $this->relation->deleteLink($this->srcAtom, new Atom(null, $this->tgtConcept), $this->relationIsFlipped);
+	        if(is_null($value)) $this->relation()->deleteLink($this->srcAtom, new Atom(null, $this->tgtConcept), $this->relationIsFlipped);
 	        // Replace by other atom => addLink
-	        else $this->relation->addLink($this->srcAtom, new Atom($value, $this->tgtConcept), $this->relationIsFlipped);
+	        else $this->relation()->addLink($this->srcAtom, new Atom($value, $this->tgtConcept), $this->relationIsFlipped);
 	        
 	    }else{
 	        throw new Exception ("Unknown patch replace. Please contact the application administrator", 500);
@@ -665,6 +723,7 @@ class InterfaceObject {
 	public function doPatchAdd($patch){
 	    // CRUD check
 	    if(!$this->crudU) throw new Exception("Update is not allowed for path '{$this->path}'", 403);
+        if($this->isRef()) throw new Exception ("Cannot update on reference interface in '{$this->path}'. See #498", 501);
 	    
 	    // Check if patch value is provided
 	    if(!array_key_exists('value', $patch)) throw new Exception ("Cannot patch add. No 'value' specfied in '{$this->path}'", 400);
@@ -678,17 +737,22 @@ class InterfaceObject {
 	
 	    // Interface is a relation to an object
 	    }elseif($this->tgtConcept->isObject){
-	        // Check: If tgtAtom (value) does not exists and there is not crud create right, throw exception
-	        if(!$this->crudC && !$tgtAtom->atomExists()) throw new Exception ("Resource '{$tgtAtom->__toString()}' does not exist and may not be created in {$this->path}", 403);
-	        	
-	        $this->relation->addLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
-	
+	        // Check if atom exists and may be created (crudC)
+	        if(!$tgtAtom->atomExists()){
+                if($this->crudC) $tgtAtom->addAtom();
+                else throw new Exception ("Resource '{$tgtAtom->__toString()}' does not exist and may not be created in {$this->path}", 403);
+            }
+	        
+            // Add link when possible (relation is specified)	
+	        if(is_null($this->relation)) $this->logger->debug("addLink skipped because '{$this->path}' is not an editable expression");
+            else $this->relation()->addLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
+            
 	    // Interface is a relation to a scalar (i.e. not an object)
 	    }elseif(!$this->tgtConcept->isObject){    	
 	        // Check: If interface is univalent, throw exception
 	        if($this->isUni) throw new Exception("Cannot patch add for univalent interface {$this->path}. Use patch replace instead", 500);
 	        	
-	        $this->relation->addLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
+	        $this->relation()->addLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
 	        
 	    }else{
 	        throw new Exception ("Unknown patch add. Please contact the application administrator", 500);
@@ -704,6 +768,7 @@ class InterfaceObject {
 	public function doPatchRemove($patch){	   
 	    // CRUD check
 	    if(!$this->crudU) throw new Exception("Update is not allowed for path '{$this->path}'", 403);
+        if($this->isRef()) throw new Exception ("Cannot update on reference interface in '{$this->path}'. See #498", 501);
 	    
         // Check if patch value is provided
 	    if(!array_key_exists('value', $patch)) throw new Exception ("Cannot patch remove. No 'value' specfied in '{$this->path}'", 400);
@@ -718,13 +783,13 @@ class InterfaceObject {
 		// Interface is a relation to an object
         }elseif($this->tgtConcept->isObject){
 			
-			$this->relation->deleteLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
+			$this->relation()->deleteLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
 		
 		// Interface is a relation to a scalar (i.e. not an object)
         }elseif(!$this->tgtConcept->isObject){
 			if($this->isUni) throw new Exception("Cannot patch remove for univalent interface {$this->path}. Use patch replace instead", 500);
 			
-			$this->relation->deleteLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
+			$this->relation()->deleteLink($this->srcAtom, $tgtAtom, $this->relationIsFlipped);
 			
 		}else{
 			throw new Exception ("Unknown patch remove. Please contact the application administrator", 500);
@@ -778,15 +843,13 @@ class InterfaceObject {
 	}
 	
 	/**
-	 * Returns all toplevel interface objects that are not assigned to a role
+	 * Returns all toplevel interface objects that are public (i.e. not assigned to a role)
 	 * @return InterfaceObject[]
 	 */
 	public static function getPublicInterfaces(){
-	    $interfaces = array();
-	    foreach(InterfaceObject::getAllInterfaces() as $ifc){
-	        if (empty($ifc->ifcRoleNames)) $interfaces[] = $ifc;
-	    }
-	    return $interfaces;
+	    return array_values(array_filter(InterfaceObject::getAllInterfaces(), function($ifc){
+            return $ifc->isPublic();
+        }));
 	}
 	
 	/**
@@ -802,10 +865,9 @@ class InterfaceObject {
 	    
 	    
 	    foreach ($allInterfaceDefs as $ifcDef){
-	        $ifc = new InterfaceObject($ifcDef['ifcObject']);
+	        $ifc = new InterfaceObject($ifcDef['ifcObject'], null, true);
 	        
 	        // Set additional information about this toplevel interface object
-	        $ifc->isTopLevelIfc = true; // Specify as top level ifc
 	        $ifc->ifcRoleNames = $ifcDef['interfaceRoles'];
 	        
 	        self::$allInterfaces[$ifc->id] = $ifc;